+++ backend/feeder/mixins.py
@@ -47,48 +47,3 @@ class CommentMixin(models.Model):
 
     class Meta:
         abstract = True
-
-
-class SaveHistoryDataViewSetMixin:
-    def perform_create(self, serializer):
-        user = self.request.user
-        if hasattr(user, "uuid"):
-            request_user = user.uuid
-        else:
-            raise ValidationError({"permission": "You do not have permissions to perform this action"})
-
-        super().perform_create(serializer)
-
-        instance = serializer.instance
-        History().entry_creation(status=History.STATUS_CREATE, instance=instance, request_user_uuid=request_user)
-
-    def perform_update(self, serializer):
-        user = self.request.user
-        if hasattr(user, "uuid"):
-            request_user = user.uuid
-        else:
-            raise ValidationError({"permission": "You do not have permissions to perform this action"})
-
-        super().perform_update(serializer)
-
-        instance = serializer.instance
-        History().entry_creation(status=History.STATUS_UPDATE, instance=instance, request_user_uuid=request_user)
-
-    def perform_destroy(self, instance):
-        user = self.request.user
-        if hasattr(user, "uuid"):
-            request_user = user.uuid
-        else:
-            raise ValidationError({"permission": "You do not have permissions to perform this action"})
-
-        instance_id = instance.id
-
-        super().perform_destroy(instance)
-
-        History().entry_creation(
-            status=History.STATUS_DELETE,
-            instance=instance,
-            request_user_uuid=request_user,
-            instance_id=instance_id
-        )
-
