+++ backend/feeder/models.py
@@ -4,6 +4,7 @@ from django.core.exceptions import ValidationError
 from django.utils.translation import gettext_lazy as _
 
 from feeder.mixins import TimeMixin, CommentMixin, NameMixin
+from feeder.soft_delete import SoftDeleteModelMixin
 
 
 def gen_uuid():
@@ -103,7 +104,7 @@ class Engagement(TimeMixin):
     notion_id = models.CharField(max_length=255, db_index=True)
 
 
-class Volunteer(TimeMixin):
+class Volunteer(TimeMixin, SoftDeleteModelMixin):
     """ Волонтеры """
     uuid = models.UUIDField(default=gen_uuid, unique=True, db_index=True)
     person = models.ForeignKey(Person, on_delete=models.SET_NULL, null=True, blank=True)
