+++ backend/feeder/views/volunteer.py
@@ -3,7 +3,7 @@ from django_filters import rest_framework as django_filters
 from django_filters.rest_framework import DjangoFilterBackend
 
 from feeder import serializers, models
-from feeder.mixins import SaveHistoryDataViewSetMixin
+from feeder.views.mixins import SaveHistoryDataViewSetMixin, SoftDeleteViewSetMixin
 
 
 class NumberInFilter(django_filters.BaseInFilter, django_filters.NumberFilter):
@@ -59,7 +59,7 @@ class VolunteerCustomFieldValueFilter(django_filters.FilterSet):
         fields = ['custom_field', 'volunteer']
 
 
-class VolunteerViewSet(SaveHistoryDataViewSetMixin, viewsets.ModelViewSet):
+class VolunteerViewSet(SaveHistoryDataViewSetMixin, SoftDeleteViewSetMixin, viewsets.ModelViewSet):
     permission_classes = [permissions.IsAuthenticated, ]
     queryset = models.Volunteer.objects.all()
     serializer_class = serializers.VolunteerSerializer
